pyproject.toml
[project]
name = "workshop1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.13",
    "uvicorn>=0.41.0",
]
